Add migrate/initialize/destroy accepting a Session.
This is useful for apps embedding pillar as a library, so that they can
provide their already existing session.

Adds library acceptance specs for migrator+session extensions, while
not each scenario is duplicated for the session interface, for each
feature it's just one scenario testing the interface taking a session.

String.format is replaced with string interpolation as it's more readable.